Typos and my promise
Fixed a few typos i saw while reading the Read-Me.
Currently looking over my library to see what can be added or optimized.
Due to making backwards support my #1 priority some changes that I want to make to the loop and step objects will be tricky.
The idea is to change the callback method to give the self object first then the position/index for steps and make loops callback to give the self then the time value. To make this work the user would need to import those new object separately. This though opens a new can of worms though... I'll think about it...