Changed how steps and loops work
Added multi.compat.backwards[ver]
Require this to be able to use features that were changed from that version.
loops callback args are now: self,dt
steps callback args are now: self,pos
If requiring the backwards compat for the previous version do so after requiring all of the modules you are using.